api/requirements.txt
sanic==20.12.3 #21.6.2
sanic-openapi==21.6.1
PyJWT==1.7.1